main chatbot working
The main chatbot to work with the VA Loan books is currently working. I removed test documents that I was working with before getting it working. Next steps would be to add more conversation history and to re-write the question with previous context. Its also not clear how many documents its retrieving. Could clear that up a bit more.